# app/api/menu-engineering/openapi.yaml
openapi: 3.0.3
info:
  title: Carmen ERP - Menu Engineering API
  description: |
    Comprehensive Menu Engineering API for sales data integration, performance analysis, 
    and recipe optimization using Boston Matrix methodology.
    
    ## Features
    - **POS Integration**: Import and synchronize sales data from Point of Sale systems
    - **Menu Analysis**: Comprehensive performance analysis using Boston Matrix (Stars, Plowhorses, Puzzles, Dogs)
    - **AI Recommendations**: Intelligent optimization suggestions based on performance data
    - **Real-time Costing**: Dynamic recipe cost calculations with current ingredient prices
    - **Cost Monitoring**: Active alerts and threshold monitoring for cost variations
    - **Performance Metrics**: Detailed analytics for recipe performance and trends
    
    ## Security
    All endpoints require authentication via JWT or Keycloak tokens and implement:
    - Role-based access control (RBAC)
    - Rate limiting and DDoS protection
    - Input validation and sanitization
    - Comprehensive audit logging
    - Security headers and CORS policies
    
    ## Rate Limits
    - Sales Import: 50 requests/hour
    - POS Sync: 20 requests/hour
    - Analytics: 100-300 requests/hour
    - Recipe APIs: 200-500 requests/hour
    
  version: 1.0.0
  termsOfService: https://carmen-erp.com/terms
  contact:
    name: Carmen ERP API Support
    url: https://carmen-erp.com/support
    email: api-support@carmen-erp.com
  license:
    name: MIT
...
servers:
...
security:
  - BearerAuth: []
  - KeycloakAuth: []
paths:
...
  /recipes/{id}/real-time-cost:
    get:
      tags:
        - Recipe Enhancement
      summary: Get real-time recipe cost calculation
      description: |
        Calculate current recipe costs using real-time ingredient prices with comprehensive breakdown.
      operationId: getRealTimeCost
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Recipe identifier
        - name: quantity
          in: query
          schema:
            type: number
            minimum: 0.1
            default: 1
          description: Recipe quantity multiplier
        - name: currency
          in: query
          schema:
            type: string
            pattern: '^[A-Z]{3}$'
            default: USD
          description: Currency code for pricing
        - name: includeLabor
          in: query
          schema:
            type: boolean
            default: true
          description: Include labor costs in calculation
        - name: includeOverhead
          in: query
          schema:
            type: boolean
            default: true
          description: Include overhead costs in calculation
        - name: useLatestPrices
          in: query
          schema:
            type: boolean
            default: true
          description: Use latest ingredient prices
        - name: includeIngredientBreakdown
          in: query
          schema:
            type: boolean
            default: true
          description: Include detailed ingredient cost breakdown
        - name: includeAlternatives
          in: query
          schema:
            type: boolean
            default: false
          description: Include alternative supplier options
        - name: compareToTarget
          in: query
          schema:
            type: boolean
            default: false
          description: Compare actual cost to target cost
      responses:
        '200':
          description: Real-time cost calculated successfully
          headers:
            X-Recipe-Id:
              schema:
                type: string
              description: Recipe identifier
            X-Calculation-Id:
              schema:
                type: string
              description: Unique calculation identifier
            X-Total-Cost:
              schema:
                type: string
              description: Total calculated cost
            X-Gross-Margin:
              schema:
                type: string
              description: Gross margin percentage
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/RealTimeCostResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication
    KeycloakAuth:
      type: oauth2
      description: Keycloak OAuth2 authentication
      flows:
        authorizationCode:
          authorizationUrl: https://keycloak.carmen-erp.com/auth/realms/carmen/protocol/openid-connect/auth
          tokenUrl: https://keycloak.carmen-erp.com/auth/realms/carmen/protocol/openid-connect/token
          scopes:
            read:menu_engineering: Read menu engineering data
            write:menu_engineering: Write menu engineering data
            read:recipes: Read recipe data
            write:recipes: Write recipe data
  schemas:
...
    ApiResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          description: Indicates if the request was successful
        data:
          description: Response data (present on success)
        error:
          type: string
          description: Error message (present on failure)
        message:
          type: string
          description: Additional message or context
        details:
          description: Additional error details or validation errors
        metadata:
          $ref: '#/components/schemas/ApiMetadata'
        pagination:
          $ref: '#/components/schemas/PaginationInfo'
    ApiMetadata:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
        version:
          type: string
        cacheTtl:
          type: integer
          description: Cache TTL in seconds
        generatedAt:
          type: string
          format: date-time
    PaginationInfo:
      type: object
      required:
        - page
        - limit
        - total
        - totalPages
        - hasNext
        - hasPrev
      properties:
        page:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
        hasNext:
          type: boolean
        hasPrev:
          type: boolean
    Money:
      type: object
      required:
        - amount
        - currency
      properties:
        amount:
          type: number
          format: decimal
          description: Monetary amount
        currency:
          type: string
          pattern: '^[A-Z]{3}$'
          description: ISO 4217 currency code
...
    RealTimeCostResponse:
      type: object
      properties:
        recipeId:
          type: string
          format: uuid
        recipeName:
          type: string
        recipeCode:
          type: string
        calculatedAt:
          type: string
          format: date-time
        calculationId:
          type: string
        parameters:
          type: object
        costBreakdown:
          type: object
          properties:
            ingredients:
              type: array
              items:
                type: object
                properties:
                  ingredientId:
                    type: string
                  ingredientName:
                    type: string
                  quantity:
                    type: number
                  unit:
                    type: string
                  unitCost:
                    $ref: '#/components/schemas/Money'
                  totalCost:
                    $ref: '#/components/schemas/Money'
                  vendor:
                    type: string
                  priceDate:
                    type: string
                    format: date-time
                  variance:
                    type: object
                    properties:
                      amount:
                        type: number
                      currency:
                        type: string
                      percentage:
                        type: number
                  alternatives:
                    type: array
                    items:
                      type: object
                      properties:
                        vendor:
                          type: string
                        unitCost:
                          $ref: '#/components/schemas/Money'
            totals:
              type: object
              properties:
                ingredientCost:
                  $ref: '#/components/schemas/Money'
                laborCost:
                  $ref: '#/components/schemas/Money'
                overheadCost:
                  $ref: '#/components/schemas/Money'
                totalDirectCost:
                  $ref: '#/components/schemas/Money'
                costPerUnit:
                  $ref: '#/components/schemas/Money'
        pricing:
          type: object
          properties:
            currentSellingPrice:
              $ref: '#/components/schemas/Money'
            suggestedSellingPrice:
              $ref: '#/components/schemas/Money'
            grossMargin:
              $ref: '#/components/schemas/Money'
            grossMarginPercentage:
              type: number
        alerts:
          type: array
          items:
            type: object
...
  responses:
    BadRequest:
      description: Bad Request - Invalid input parameters
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Invalid request parameters"
    Unauthorized:
      description: Unauthorized - Authentication required
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Authentication required"
...
    NotFound:
      description: Not Found - Resource does not exist
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Resource not found"
...
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Internal server error"
...